+++ docs/Changelog.md
@@ -32,6 +32,8 @@ This page is a brief overview of each version.
 
 ### Misc
 - Update system prompt to include basic SB formatting hints and docs links.
+- Add support for structured responses using json schema
+- Refactor several commands to space lua instead of typescript
 
 ---
 ## 0.5.0 (2025-12-15)